Added module dependency support.
In order for this to work you need to arrange for a modules.dep file to
exist in the same place as your modules.  To create this file, cd to the
module directory and run depmod.pl from there.  Note that the default
Makefile will do this for you on "make install".

Things should fallback gracefully if you lack a modules.dep file (but then
you won't get the benefit of course).

insmod() now automatically loads needed modules first.
rmmod() now automatically removes dependent modules first.
New functions preparereload() and reloadmarked() can be used either side of
a reload to mark dependent modules which are to be removed, and then reload
them, respectively.  control and noperserv's reload methods have been
updated to use these as appropriate.