Added support for remembering the oldest-ever timestamp seen for a channel. Added to the database with a function to change the timestamp entry on a channel when needed.
This timestamp is then used when we need to join a channel - the
localburstontochannel() function is used with that timestamp as an argument.
This means we can get the ircd to enforce the sanctity of +i/+k channels via
the netrider code, and also take care of rogue topics, ops, bans and the
like.

Also two unrelated fixes - one to save when we automatically remove +k from
forcemodes because no key was set, and another to prevent the wrong
maxuserid being recorded during load.