Added "nsmalloc" to the core.
This is a simple malloc() wrapper which supports "pools".  It is possible to
free all allocations belonging to a given pool at the same time - perfect
for stopping modules from leaking memory.

If you want to use this in your module, add a #define to nsmalloc.h to grab
a number.