Chanserv refactoring
The main change is that the database has been pulled into a separate module.
It now uses the psql module for database services.  With the other
refactoring changes to the newserv core, the chanservdb module is now
completely independent of the irc module - so the entire IRC stack can be
restarted without reloading the chanserv database.

Moved the database items into a subdirectory to make things tidier - other
sections will follow in time.

Also changed the top-level Makefile to support modules with their own
specific clean requirements - just add a "clean" target to the
subdirectory's Makefile and then add it to the CLEANDIRS list in the
top-level Makefile.