More chanserv refactoring
The previously monolithic "chancmds.c", "usercmds.c", and "authcmds.c" files
have been split up into multiple files in new directories "chancmds",
"usercmds" and "authcmds".

Each command source file now has a comment header which contains tags
indicating the name of the command, the level required to use it, the
maximum argument count, a short description and the name of the command
function to be called.  A perl script has been added which reads these tags
and automatically generates a file called "commandlist.c" which sets up
appropriate _init() and _fini() functions for the module.  The perl script
also generates the Makefile, so if you add a command it will be added to the
Makefile automatically.  The perl script will emit warnings if tags are
incomplete and not include that file.

A perl script called "refactor.pl" has been added which assists in the
splitting up process.  This was used to generate the new individual files in
each directory.  You may find it useful if you have locally added your own
commands to one of the old monolithic files and want it to automatically
generate new individual files for you.  Note it's not perfect, in particular
it won't notice if you have defined helper functions as well as your main
command function, so don't overwrite the checked-in version of files with
ones generated by refactor.pl unless you have made changes to the relevant
command and have checked it first.

If you want to register command aliases, create a dummy .c file with
appropriate tags (see authcmds/auth.c for an example).