Added CIDR ban support.
unsigned ints for mask and ipaddr have been added to the chanban structure.
If a CIDR or exact IP ban is detected, these are filled in and subsequently
used during ban matching.

A fix to channel.c to enable "reload channel" to work has also crept in, as
well as a 4-line rewrite of some retarded code in channel.c