Added capability to pqsql module to use default unix socket.
Updated pqsql so that if a host of "UNIX" is specified it will not pass
hostname/port information to the connect call, causing the postgres library
to fall back on the default unix domain socket.  This avoids the need to
specify the unix socket path in the newserv config file - desireable since
it can vary depending on the postgres installation.

The comparison is case sensitive so if your DB host is really called "unix"
then specifying the name in lowercase will cause it to connect via tcp as
before.

Also made "UNIX" the default hostname.

Also included the strlfunc.h to remove a compiler warning.