Tightened up the auth code a bit.
The user limit check is now done before logging last user@host etc.
Messages are now sent to other account users when someone auths - and a different message sent if the account is at the auth limit.
Fixed up a couple of other incorrect error messages (in removeuser.c).
chancmds/commandlist.c seems to be a changeset or two behind so there's a new one of those too.