Added some things to chanserv_newsearch.
Channel display format "qusers" displays a summary of the known users on that channel.
Channel search term "qusers" returns the number of known users on that channel.  Optional argument is the flags the user must have (all flags specified with + and none with - must be present for it to match).