Add perfect forward secrecy to nterfacer
This bumps the nterfacer protocol version to 5.

The authentication stage of the protocol is modified. In addition to the
challenge (from server) or response (from client), the IV and the nonce,
both sides send a Curve25519 public key in hex. The new field is added
to the end of the challenge/response + IV + nonce line.

The resulting shared secret from X25519 is then hashed like so:
SHA-256(shared from smult || receiver public key || sender public key).
The order for hashing was taken from
https://download.libsodium.org/doc/advanced/scalar_multiplication.html.
The resulting hash is then used as the client raw key, meaning that the
nterfacer rekeying mechanism/subkeying mechanism is untouched from
version 4.

This adds a Curve25519 implementation (curve25519-donna) to /lib/.
Deploying this changeset therefore requires recompiling and  restarting
newserv. curve25519.c contains helper functions to avoid misuse of the
underlying ECDH primitive.